Fix issues

+++ urban-spork-common/src/com/urbanspork/common/protocol/vmess/header/SecurityType.java
@@ -6,25 +6,25 @@ import java.util.Arrays;
 
 public enum SecurityType {
 
-    UNKNOWN(0),
-    LEGACY(1),
-    AUTO(2),
-    AES128_GCM(3),
-    CHACHA20_POLY1305(4),
-    NONE(5),
-    ZERO(6);
+    UNKNOWN((byte) 0),
+    LEGACY((byte) 1),
+    AUTO((byte) 2),
+    AES128_GCM((byte) 3),
+    CHACHA20_POLY1305((byte) 4),
+    NONE((byte) 5),
+    ZERO((byte) 6);
 
-    private final int value;
+    private final byte value;
 
-    SecurityType(int value) {
+    SecurityType(byte value) {
         this.value = value;
     }
 
-    public int getValue() {
+    public byte getValue() {
         return value;
     }
 
-    public static SecurityType valueOf(int value) {
+    public static SecurityType valueOf(byte value) {
         return Arrays.stream(values()).filter(type -> type.value == value).findFirst().orElse(UNKNOWN);
     }
